Прибрался в проекте, починил баг в меню, реализовал фон в отдельную функцию, все страницы приведены к базовому виду.

+++ lib/7_PageManager/PageManager.h
@@ -11,5 +11,5 @@ class PageManager {
         Page* getCurrent();
     private:
         std::vector<Page*> _pages;
-        uint8_t _currentIndex = 0;
+        int _currentIndex = 0;
 };
